% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_df_folds.R
\name{calculate_df_folds}
\alias{calculate_df_folds}
\title{Split a given data frame into folds for building confidence intervals}
\usage{
calculate_df_folds(df, train, test)
}
\arguments{
\item{df}{The data frame to be used to split into folds.}

\item{train}{Number of training weeks.}

\item{test}{Number of test weeks.}
}
\value{
Test.
}
\description{
A detailed description of what the function does.
}
\examples{
# Examples of how to use the function:
data <- data.frame(week = 1:10, sales = runif(10, 100, 1000))
result <- calculate_df_folds(data, 2, 2)
result
}
