% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rolling_regression.R
\name{fit_rolling_regression}
\alias{fit_rolling_regression}
\title{Title}
\usage{
fit_rolling_regression(df_list, include_intercept = FALSE, model_type = "OLS")
}
\arguments{
\item{model_type}{}
}
\description{
Title
}
